move auto task trigger to refreshVillage instead of update village
